fix broken test amidst.mojangapi.world.test.TestWorldEntryEquality
